Database File now works, added database tester main
I tweaked the DbConnections file so now it WORKS! It might be a little confusing to use. Ask Dallin (me) if you have questions about it :D

You will need to make a unique method for every time you want to call a SelectCommand, which is the command that prints a row of the table.

I also added a new main called BaseTesting for testing the Database connectivity